% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtools_aggregation.R
\name{sigtools_aggregation}
\alias{sigtools_aggregation}
\title{Generates aggregation plots for all the signals in the given multi-column bedGraph file, over the specified elements in the given interval file}
\usage{
sigtools_aggregation(
  path_mulColBedG,
  path_reg,
  chr,
  mode,
  header = FALSE,
  prefix = "s",
  sep = "\\t",
  outdir = "aggregation",
  resolution = 200L,
  neighborhood = 50L,
  percentile = 100,
  enrichment = FALSE,
  means = c(NA)
)
}
\arguments{
\item{path_mulColBedG}{The path to the input multi-column bedgraph file}

\item{chr}{The intended chromosome (example:21L)}

\item{mode}{Controls the length unification process [options: 'point' or 'stretch']}

\item{header}{Does your file include headers? [default:FALSE]}

\item{prefix}{Your preferred prefix for naming the signals [default:'s']}

\item{sep}{Your field separator character [default:'\t']}

\item{outdir}{Name of the output directory [default:aggregation]}

\item{resolution}{The bin size of the input file [default:200]}

\item{neighborhood}{The number of covered up and down stream bins [default:50]}

\item{percentile}{What percentile of your data you want to work with [default:'100']}

\item{enrichment}{Do you like to plot the enrichment? [default:FALSE]}

\item{means}{If given, istead of calculating mean values of the signals, the mean will be obtained from this list [default:c(NA)]}

\item{path_regThe}{The path to the inpput interval file}
}
\value{

}
\description{
Generates aggregation plots for all the signals in the given multi-column bedGraph file, over the specified elements in the given interval file
}
\examples{
sigtools_aggregation(path_mulColBedG = "./test_data/E003-assays_chr21_bin200_bwtool.mcBedGraph", path_reg = "./test_data/Ensembl_chr21.gene_info", chr = 21, mode = 'stretch', header = TRUE , outdir = 'aggregation_E003-chr21_enriched_2', resolution = 200L , neighborhood = 10L, enrichment = TRUE, means = c(0.640, 0.601, 0.633, 0.575, 0.604, 0.597) )
}
