% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtools_autocorrelation.R
\name{sigtools_autocorrelation}
\alias{sigtools_autocorrelation}
\title{Plots the autocorrelation of several signals}
\usage{
sigtools_autocorrelation(
  path_mulColBedG,
  outdir = "autocorrelation",
  header = FALSE,
  prefix = "s",
  sep = "\\t",
  resolution = 200L,
  lag = 50L,
  mode = "bins",
  x_label = NA,
  y_label = NA,
  img_title = NA,
  img_width = 10,
  img_height = 10,
  fontSize = 20
)
}
\arguments{
\item{path_mulColBedG}{The path to the input file of multi-column bedgraph}

\item{outdir}{Name of the output directory [default:autocorrelation]}

\item{header}{Does your file include headers? [default:FALSE]}

\item{prefix}{Your preferred prefix for naming the signals [default:'s']}

\item{sep}{Your field separator character [default:'\t']}

\item{resolution}{The bin size of the input file [default:200]}

\item{lag}{The number of demanded shifts [default:50]}

\item{mode}{If set to 'regions', the autocorrelation is calculated over adjacent elements mentioned in the additional bed file [default:bins] [options:'bins' or 'regions']}

\item{x_label}{Label for the X axis}

\item{y_label}{label for the Y axis}

\item{img_title}{Title of the graph}

\item{img_width}{Width of the saved image (cm) [default:10]}

\item{img_height}{Height of the saved image (cm) [default:10]}

\item{fontSize}{The font size of the saved image [default:20]}
}
\value{

}
\description{
Plots the autocorrelation of several signals
}
\examples{
sigtools_autocorrelation("./test_data/E003-assays_chr21_bin200_bwtool.mcBedGraph", outdir = "autocorrelation", header = TRUE, prefix = 's', sep = '\t', resolution = 200L, lag = 50L, img_height = 10, img_width = 20)
}
