% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtools_concatenation.R
\name{sigtools_concatenate}
\alias{sigtools_concatenate}
\title{Title}
\usage{
sigtools_concatenate(
  path,
  pattern,
  num_signals,
  header = FALSE,
  sep = "\\t",
  outdir = "./preprocessing",
  sampling = TRUE,
  population_per = 0.1,
  population_ratio = c(1, 2)
)
}
\arguments{
\item{path}{The path to the directory containing your cell types}

\item{pattern}{The pattern in which you named your cell types}

\item{num_signals}{The number of signals presented in the dataset}

\item{header}{Does this file include headers? [default: FALSE]}

\item{sep}{The field separator character [default: '\t']}

\item{outdir}{The name of the output file [default: 'preprocessing']}

\item{sampling}{Wheather you want the input file(s) to be sampled? [default: TRUE]}

\item{population_per}{The size of the sample relative to the size of the dataset  [default: 0.01]}

\item{population_ratio}{The number of sampled intervals relative to their length [default: c(1,2)]}
}
\value{

}
\description{
Title
}
\examples{
sigtools_concatenate('../compCan/ssmEdited', "exp73-7-E*_edited", 3)
}
