% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtools_correlation.R
\name{sigtools_correlation}
\alias{sigtools_correlation}
\title{Plots a heatmap of pair-wise correlation of sets of genomic signals}
\usage{
sigtools_correlation(
  path_mulColBedG_1,
  header_1 = FALSE,
  prefix_1 = "set1",
  path_mulColBedG_2,
  header_2 = FALSE,
  prefix_2 = "set2",
  outdir = "correlation",
  sep = "\\t",
  enrichment = FALSE,
  x_label = NA,
  y_label = NA,
  img_title = NA,
  img_width = 10,
  img_height = 10,
  fontSize = 20
)
}
\arguments{
\item{path_mulColBedG_1}{The path to the file of multi-column bedgraph to be considered as the first set of signals}

\item{header_1}{Does your file include headers? [default:FALSE]}

\item{prefix_1}{Your preferred prefix for naming the signals in the first set [default:'set1_']}

\item{path_mulColBedG_2}{The path to the file of multi-column bedgraph to be considered as the second set of signals}

\item{header_2}{Does your file include headers? [default:FALSE]}

\item{prefix_2}{Your preferred prefix for naming the signals in the second set [default:'set2_']}

\item{outdir}{Name of the output directory [default:correlation]}

\item{sep}{Your field separator character [default:'\t']}

\item{enrichment}{Do you like to plot the enrichment? [default:FALSE]}

\item{x_label}{Label for the X axis}

\item{y_label}{label for the Y axis}

\item{img_title}{Title of the graph}

\item{img_width}{Width of the saved image (cm) [default:10]}

\item{img_height}{Height of the saved image (cm) [default:10]}

\item{fontSize}{The font size of the saved image [default:20]}
}
\value{

}
\description{
Plots a heatmap of pair-wise correlation of sets of genomic signals
}
\examples{
sigtools_correlation(path_mulColBedG_1 = "./test_data/E003-assays_chr21_bin200_bwtool.mcBedGraph", header_1 = TRUE, path_mulColBedG_2 = "./test_data/E003-assays_chr21_bin200_bwtool.mcBedGraph", header_2 = TRUE, outdir = "correlation", img_width = 15, img_height = 15)
}
