% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtools_distribution.R
\name{sigtools_distribution}
\alias{sigtools_distribution}
\title{Plots the distribution of several signals}
\usage{
sigtools_distribution(
  path_mulColBedG,
  outdir = "distribution",
  plots = c("curve", "ecdf", "box"),
  header = FALSE,
  prefix = "s",
  sep = "\\t",
  percentile = 100,
  enrichment = FALSE,
  nozero = FALSE,
  x_label = NA,
  y_label = NA,
  img_title = NA,
  img_width = 10,
  img_height = 10,
  fontSize = 20
)
}
\arguments{
\item{path_mulColBedG}{The path to the input file of multi-column bedgraph to depict distribution for}

\item{outdir}{Name of the output directory [default:distribution]}

\item{plots}{Options of distribution plots [default:plots = c("curve", "ecdf", "box")]}

\item{header}{Does your file include headers? [default:FALSE]}

\item{prefix}{Your preferred prefix for naming the signals [default:'s']}

\item{sep}{Your field separator character [default:'\t']}

\item{percentile}{What percentile of your data you want to work with [default:'100']}

\item{enrichment}{Do you like to plot the enrichment? [default:FALSE]}

\item{nozero}{Do you want to remove the zero entries? [default:FALSE]}

\item{x_label}{Label for the X axis}

\item{y_label}{label for the Y axis}

\item{img_title}{Title of the graph}

\item{img_width}{Width of the saved image (cm) [default:10]}

\item{img_height}{Height of the saved image (cm) [default:10]}

\item{fontSize}{The font size of the saved image [default:20]}
}
\value{

}
\description{
Plots the distribution of several signals
}
\examples{
sigtools_distribution("./test_data/E003-assays_chr21_bin200_bwtool.mcBedGraph", plots = c("curve"), header = TRUE, percentile = 99, enrichment = FALSE, nozero = TRUE, img_height = 10, img_width = 20, y_label = "")
}
